produce/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
h11>=0.13,<0.15
httpx==0.25.0
idna==3.10
kafka-python==2.0.2
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.3
typing-extensions==4.15.0
uvicorn==0.35.0
